Assignment 3/Streamlit/Spotify/requirements.txt
absl-py==0.11.0
alembic==1.4.2
altair==4.1.0
annoy==1.17.0
apispec==1.3.3
argcomplete==1.11.1
argon2-cffi==20.1.0
asgiref==3.2.7
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
awscli==1.18.82
Babel==2.8.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.33
botocore==1.17.33
cached-property==1.5.1
cachetools==4.1.1
cattrs==0.9.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.34
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
Django==3.0.6
dnspython==1.16.0
docutils==0.15.2
elasticsearch==7.9.1
email-validator==1.1.1
entrypoints==0.3
enum-compat==0.0.3
faiss-cpu==1.6.4
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.14
grpcio==1.33.2
gunicorn==19.10.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.17.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.5.0
lockfile==0.12.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.2
mistune==0.8.4
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pendulum==1.4.4
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.12.0
prison==0.1.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
py4j==0.10.9
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pyspark==3.0.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==3.0.4
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
sqlparse==0.3.1
streamlit==0.70.0
tabulate==0.8.7
tenacity==4.12.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
threadpoolctl==2.1.0
thrift==0.13.0
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.51.0
traitlets==4.3.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.9
validators==0.18.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
zope.deprecation==4.4.0
